+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## [TBD]
+* Removed deprecated methods from native auth public interface (#2588)
+
 ## [1.9.0]
 * Add feature flags provider to be controlled from broker (#2540)
 * Added GitHub issue templates for better issue tracking and reporting (#2554)